flask_app/requirements.txt
Flask
psycopg2-binary
matplotlib
